Extend HwROI to Pilatus2 6M detector.
Current implementation of the Hardware ROI management for
Pilatus3 is extended to Pilatus2 6M model. To restrict this
extension to the 6M model, the m_has_harware_roi flag is
initialized to true and is set to false for detector models
 which could not be tested i.e. Pilatus2 2M and 1M models.
The addition of those  excluded models would be straigh forward.

Just the extension to deal with Pilatus2 6M is not enough. We
observe that Pilatus2 6M detector goes to ERROR after sending
the setroi command. However, this return code from the camserver
looks a fake, since this was an experimental implementation
(according to Dectris) and, in fact, the roi is properly set.
The workaround implemented here ignores this string error and
sets the detector state to STANBY ONLY for Pilatus2 series,
because this behavior has been only observed for Pilatus2 6M
model using the tvx-7.3.13-121212 version.